python bindings (NOTE: read full message!)
For final deployment we want all the C++ implementations to use floats,
but it is hard to make our finite difference derivative checks work with
them. So I guess the plan is to implement all the bindings with doubles,
then figure out some way to compare the final results when switching to
floats without checking all the intermediate derivatives.